requirements.txt
flask
requests
airtable-python-wrapper
